Move current "runner framework" to overlay-runner implementation
